requirements.txt
numpy
pygame>=2.5.0
PyOpenGL>=3.1.7
gymnasium>=0.29.0
stable-baselines3>=2.3.0
torch>=2.1.0
tensorboard>=2.15.0
